Moved FT's to their individual files and classes
