+++ functional_tests/test_layout_and_styling.py
@@ -1,0 +1,21 @@
+from .base import FunctionalTest
+
+
+# CSS testcases
+class LayoutAndStylingTest(FunctionalTest):
+    def test_layout_styling(self):
+        self.browser.get(self.live_server_url)
+        self.browser.set_window_size(1024, 768)
+
+        inputbox = self.browser.find_element_by_id('id_new_item')
+        self.assertAlmostEqual(
+            inputbox.location['x'] + inputbox.size['width'] / 2,
+            512, delta=5
+        )
+        inputbox.send_keys('testing\n')
+        inputbox = self.browser.find_element_by_id('id_new_item')
+        self.assertAlmostEqual(
+            inputbox.location['x'] + inputbox.size['width'] / 2,
+            512,
+            delta=5
+            )
