+++ functional_tests/test_list_item_validation.py
@@ -1,0 +1,7 @@
+from unittest import skip
+from .base import FunctionalTest
+
+
+class ItemValidationTest(FunctionalTest):
+    def test_cannot_empty_list_of_items(self):
+        self.fail('write me!')
