+++ functional_tests/test_simple_list_creation.py
@@ -1,45 +1,10 @@
-import os
+from .base import FunctionalTest
 from selenium import webdriver
 from selenium.webdriver.common.keys import Keys
-import unittest
-from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 
 
 # Testcase class
-class NewVisitorTest(StaticLiveServerTestCase):
-
-    # setup the chrome driver
-    def setUp(self):
-        self.chromedriver = "~/Downloads/chromedriver"
-        os.environ["webdriver.chrome.driver"] = self.chromedriver
-        self.browser = webdriver.Chrome(self.chromedriver)
-        self.browser.implicitly_wait(3)
-
-    # Driver quit function
-    def tearDown(self):
-        self.browser.quit()
-
-    def check_if_row_exists_in_list_table(self, row_text):
-        table = self.browser.find_element_by_id('id_list_table')
-        rows = table.find_elements_by_tag_name('tr')
-        self.assertIn(row_text, [row.text for row in rows])
-
-    def test_layout_styling(self):
-        self.browser.get(self.live_server_url)
-        self.browser.set_window_size(1024, 768)
-
-        inputbox = self.browser.find_element_by_id('id_new_item')
-        self.assertAlmostEqual(
-            inputbox.location['x'] + inputbox.size['width'] / 2,
-            512, delta=5
-        )
-        inputbox.send_keys('testing\n')
-        inputbox = self.browser.find_element_by_id('id_new_item')
-        self.assertAlmostEqual(
-            inputbox.location['x'] + inputbox.size['width'] / 2,
-            512,
-            delta=5
-            )
+class NewVisitorTest(FunctionalTest):
 
     def test_can_start_a_list_and_retrieve_later(self):
         self.browser.get(self.live_server_url)
